backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1

# Vector search (optional unless SEARCH_BACKEND=vector)
chromadb==0.5.5
sentence-transformers==3.0.1
numpy==1.26.4
tqdm==4.66.5

# Postgres support (optional unless SEARCH_BACKEND=postgres)
psycopg2-binary==2.9.9
